refactor: pulled out CreatePost into separate component similar to UpdatePost
